refactor: animation componentを分離
