+++ src/components/animations/index.ts
@@ -1,0 +1,3 @@
+import { LoadingAnimation } from "./loadingAnimation"
+import { TextAnimation } from "./textAnimation"
+export { LoadingAnimation, TextAnimation }
