+++ src/components/animations/loadingAnimation/index.module.sass
