+++ src/components/index.ts
@@ -1,0 +1,2 @@
+import { LoadingAnimation, TextAnimation } from "./animations"
+export { LoadingAnimation, TextAnimation }
